파산 및 레버리지 로직 추가

+++ src/main/java/com/corinne/corinne_be/dto/transaction_dto/BuyRequestDto.java
@@ -8,13 +8,13 @@ public class BuyRequestDto {
     private String tiker;
     private int leverage;
     private int tradePrice;
-    private int buyAmount;
+    private Long buyAmount;
 
 
     public BuyRequestDto() {
     }
 
-    public BuyRequestDto(String tiker, int leverage, int tradePrice, int buyAmount) {
+    public BuyRequestDto(String tiker, int leverage, int tradePrice, Long buyAmount) {
         this.tiker = tiker;
         this.leverage = leverage;
         this.tradePrice = tradePrice;
